+++ NEWS.md
@@ -5,7 +5,8 @@
 BUGFIXES :  
 
 * `setSimulationPathAPI()` error message with bad id
-* `readInputThermal()` / `readInputRES()` fix "opts" parameter linked in sub function 
+* correction in `readClusterDesc()` calls to add "opts"
+
 
 # antaresRead 2.6.0