+++ NEWS.md
@@ -1,5 +1,12 @@
 > Copyright © 2016 RTE Réseau de transport d’électricité
 
+# antaresRead 2.6.1 (devlopment)
+
+BUGFIXES :  
+
+* `setSimulationPathAPI()` error message with bad id
+* correction in `readClusterDesc()` calls to add "opts"
+
 
 # antaresRead 2.6.0